feat/#67 : 리액트쿼리 분리
