--- src/utils/apis/fetchGroupFeed.ts
@@ -1,3 +1,0 @@
-// import { client } from './axios';
-
-// export
